+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/terrains/config/rough.py
@@ -9,7 +9,7 @@ import omni.isaac.orbit.terrains as terrain_gen
 
 from ..terrain_cfg import TerrainGeneratorCfg
 
-ASSORTED_TERRAINS_CFG = TerrainGeneratorCfg(
+ROUGH_TERRAINS_CFG = TerrainGeneratorCfg(
     size=(8.0, 8.0),
     border_width=20.0,
     num_rows=10,
@@ -50,3 +50,4 @@ ASSORTED_TERRAINS_CFG = TerrainGeneratorCfg(
         ),
     },
 )
+"""Rough terrains configuration."""